+++ CHANGELOG.md
@@ -4,6 +4,8 @@ VC Verifier Changelog
 WIP
 ---
 
+3.1.1 (2025-11-24)
+
 - DID documents and Verifiable Credentials use 60-minute TTL cache for data freshness
 - Static resources (contexts, schemas) use permanent cache